se agrega modifyRestaurants y se organiza tabla de Cards
