+++ FRONT/react-admin/src/Components/Card/Card.jsx
@@ -3,7 +3,7 @@ import style from "./Card.module.css";
 
 const Card = (props) => {
   return (
-    <div>
+    <>
       <tr key={props.key}>
         <td>{props.name}</td>
         <td>{props.menu}</td>
@@ -13,12 +13,12 @@ const Card = (props) => {
         </td>
         <td>{props.address.city}</td>
         <td>{props.address.country}</td>
-        <td>
+        <td className={style.rows}>
           <button>editar</button>
           <button>borrar</button>
         </td>
       </tr>
-    </div>
+    </>
   );
 };
