+++ FRONT/react-admin/src/Components/Card/Card.module.css
@@ -5,6 +5,22 @@ table {
   border-collapse: collapse;
 }
 
+td,
+th {
+  padding: 5px 10px;
+}
+
 table {
   width: 100%;
+}
+
+.rows {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.rows button {
+  padding: 5px 10px;
+  font-size: small;
 }
