+++ FRONT/react-admin/src/Components/ModifyRestautant.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ModifyRestaurant = () => {
+  return <div></div>;
+};
+
+export default ModifyRestaurant;
