+++ FRONT/react-admin/src/View/Home.jsx
@@ -2,8 +2,9 @@
 
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import Cards from "../Components/Cards";
+import Cards from "../Components/Cards/Cards.jsx";
 import { getAllRestaurants } from "../Redux/Actions";
+import "./Home.css";
 
 const Home = () => {
   const dispatch = useDispatch();
