+++ FRONT/react-admin/src/View/Login.jsx
@@ -1,10 +1,14 @@
-import React, { useState , useEffect} from 'react'
-import './Login.css'
-import { useDispatch } from 'react-redux';
-import { setUser } from '../Redux/Actions';
+import React, { useState, useEffect } from "react";
+import "./Login.css";
+import { useDispatch } from "react-redux";
+import { setUser } from "../Redux/Actions";
 import { initializeApp } from "firebase/app";
 import { getFirestore } from "firebase/firestore";
-import { getAuth, signInWithEmailAndPassword, fetchSignInMethodsForEmail,} from "firebase/auth";
+import {
+  getAuth,
+  signInWithEmailAndPassword,
+  fetchSignInMethodsForEmail,
+} from "firebase/auth";
 import { getStorage } from "firebase/storage";
 
 const firebaseConfig = {
@@ -13,7 +17,7 @@ const firebaseConfig = {
   projectId: "prueba-de-funciones-4b9e8",
   storageBucket: "prueba-de-funciones-4b9e8.appspot.com",
   messagingSenderId: "589197491000",
-  appId: "1:589197491000:web:34d30b1db2bbb90f631fcd"
+  appId: "1:589197491000:web:34d30b1db2bbb90f631fcd",
 };
 
 const app = initializeApp(firebaseConfig);
@@ -21,18 +25,16 @@ const auth = getAuth(app);
 const db = getFirestore(app);
 const storage = getStorage(app);
 
-
 function Login() {
   const [emailLogin, setEmailLogin] = useState("");
   const [passwordLogin, setPasswordLogin] = useState("");
-  const [user , setUser] = useState(null); //set global state
-  const [error, setError]= useState(null)
-  
+  const [user, setUser] = useState(null); //set global state
+  const [error, setError] = useState(null);
 
-//falta mantener en el estado global en useEffect
-//falta disparar la action de setUser 
-//falta identificar si es user.admin o es user.superAdmin
-//falta proteger la rutas 
+  //falta mantener en el estado global en useEffect
+  //falta disparar la action de setUser
+  //falta identificar si es user.admin o es user.superAdmin
+  //falta proteger la rutas
 
   const handleSubmit = async (event) => {
     event.preventDefault();
@@ -51,8 +53,6 @@ function Login() {
 
       setUser(userCredential.user);
       window.location.href = "/home";
-
-
     } catch (error) {
       console.error("Sign in failed!", error);
       setError(error.message);
@@ -76,6 +76,8 @@ function Login() {
         onChange={(event) => setPasswordLogin(event.target.value)}
       />
       <button type="submit">Login</button>
+
+      <span> </span>
     </form>
   );
 }
